card-generator/requirements.txt
Flask==2.3.3
reportlab==4.0.8
qrcode
pillow